@@ -1,9 +1,9 @@
-const express = require('express');
-const app = express();
-const employeeRoute = express.Router();
+const express = require('express')
+const app = express()
+const employeeRoute = express.Router()
 
 // Employee model
-let Employee = require('../models/Employee');
+let Employee = require('../models/Employee')
 
 // Add Employee
 employeeRoute.route('/create').post((req, res, next) => {
@@ -14,7 +14,7 @@ employeeRoute.route('/create').post((req, res, next) => {
       res.json(data)
     }
   })
-});
+})
 
 // Get All Employees
 employeeRoute.route('/').get((req, res) => {
@@ -38,33 +38,36 @@ employeeRoute.route('/read/:id').get((req, res) => {
   })
 })
 
-
 // Update employee
 employeeRoute.route('/update/:id').put((req, res, next) => {
-  Employee.findByIdAndUpdate(req.params.id, {
-    $set: req.body
-  }, (error, data) => {
-    if (error) {
-      return next(error);
-      console.log(error)
-    } else {
-      res.json(data)
-      console.log('Data updated successfully')
-    }
-  })
+  Employee.findByIdAndUpdate(
+    req.params.id,
+    {
+      $set: req.body,
+    },
+    (error, data) => {
+      if (error) {
+        return next(error)
+        console.log(error)
+      } else {
+        res.json(data)
+        console.log('Data updated successfully')
+      }
+    },
+  )
 })
 
 // Delete employee
 employeeRoute.route('/delete/:id').delete((req, res, next) => {
   Employee.findOneAndRemove(req.params.id, (error, data) => {
     if (error) {
-      return next(error);
+      return next(error)
     } else {
       res.status(200).json({
-        msg: data
+        msg: data,
       })
     }
   })
 })
 
-module.exports = employeeRoute;
+module.exports = employeeRoute
