@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EmployeeCreateComponent } from './employee-create.component';
 
@@ -6,12 +6,12 @@ describe('EmployeeCreateComponent', () => {
   let component: EmployeeCreateComponent;
   let fixture: ComponentFixture<EmployeeCreateComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       declarations: [ EmployeeCreateComponent ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(EmployeeCreateComponent);
