@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EmployeeEditComponent } from './employee-edit.component';
 
@@ -6,12 +6,12 @@ describe('EmployeeEditComponent', () => {
   let component: EmployeeEditComponent;
   let fixture: ComponentFixture<EmployeeEditComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       declarations: [ EmployeeEditComponent ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(EmployeeEditComponent);
