@@ -1,21 +1,20 @@
-import { Employee } from './../../model/Employee';
+import { Employee } from './../../model/employee';
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from "@angular/router";
+import { ActivatedRoute, Router } from '@angular/router';
 import { ApiService } from './../../service/api.service';
-import { FormGroup, FormBuilder, Validators } from "@angular/forms";
-
+import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-employee-edit',
   templateUrl: './employee-edit.component.html',
-  styleUrls: ['./employee-edit.component.css']
+  styleUrls: ['./employee-edit.component.css'],
 })
 
 export class EmployeeEditComponent implements OnInit {
   submitted = false;
   editForm: FormGroup;
   employeeData: Employee[];
-  EmployeeProfile: any = ['Finance', 'BDM', 'HR', 'Sales', 'Admin']
+  EmployeeProfile: any = ['Finance', 'BDM', 'HR', 'Sales', 'Admin'];
 
   constructor(
     public fb: FormBuilder,
@@ -30,17 +29,23 @@ export class EmployeeEditComponent implements OnInit {
     this.getEmployee(id);
     this.editForm = this.fb.group({
       name: ['', [Validators.required]],
-      email: ['', [Validators.required, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,3}$')]],
+      email: [
+        '',
+        [
+          Validators.required,
+          Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+.[a-z]{2,3}$'),
+        ],
+      ],
       designation: ['', [Validators.required]],
-      phoneNumber: ['', [Validators.required, Validators.pattern('^[0-9]+$')]]
-    })
+      phoneNumber: ['', [Validators.required, Validators.pattern('^[0-9]+$')]],
+    });
   }
 
   // Choose options with select-dropdown
   updateProfile(e) {
     this.editForm.get('designation').setValue(e, {
-      onlySelf: true
-    })
+      onlySelf: true,
+    });
   }
 
   // Getter to access form control
@@ -49,7 +54,7 @@ export class EmployeeEditComponent implements OnInit {
   }
 
   getEmployee(id) {
-    this.apiService.getEmployee(id).subscribe(data => {
+    this.apiService.getEmployee(id).subscribe((data) => {
       this.editForm.setValue({
         name: data['name'],
         email: data['email'],
@@ -62,10 +67,16 @@ export class EmployeeEditComponent implements OnInit {
   updateEmployee() {
     this.editForm = this.fb.group({
       name: ['', [Validators.required]],
-      email: ['', [Validators.required, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,3}$')]],
+      email: [
+        '',
+        [
+          Validators.required,
+          Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+.[a-z]{2,3}$'),
+        ],
+      ],
       designation: ['', [Validators.required]],
-      phoneNumber: ['', [Validators.required, Validators.pattern('^[0-9]+$')]]
-    })
+      phoneNumber: ['', [Validators.required, Validators.pattern('^[0-9]+$')]],
+    });
   }
 
   onSubmit() {
@@ -75,15 +86,16 @@ export class EmployeeEditComponent implements OnInit {
     } else {
       if (window.confirm('Are you sure?')) {
         let id = this.actRoute.snapshot.paramMap.get('id');
-        this.apiService.updateEmployee(id, this.editForm.value)
-          .subscribe(res => {
+        this.apiService.updateEmployee(id, this.editForm.value).subscribe({
+          complete: () => {
             this.router.navigateByUrl('/employees-list');
-            console.log('Content updated successfully!')
-          }, (error) => {
-            console.log(error)
-          })
+            console.log('Content updated successfully!');
+          },
+          error: (e) => {
+            console.log(e);
+          },
+        });
       }
     }
   }
-
 }
