@@ -4,32 +4,29 @@ import { ApiService } from './../../service/api.service';
 @Component({
   selector: 'app-employee-list',
   templateUrl: './employee-list.component.html',
-  styleUrls: ['./employee-list.component.css']
+  styleUrls: ['./employee-list.component.css'],
 })
 
 export class EmployeeListComponent implements OnInit {
-  
-  Employee:any = [];
+  Employee: any = [];
 
-  constructor(private apiService: ApiService) { 
+  constructor(private apiService: ApiService) {
     this.readEmployee();
   }
 
   ngOnInit() {}
 
-  readEmployee(){
+  readEmployee() {
     this.apiService.getEmployees().subscribe((data) => {
-     this.Employee = data;
-    })    
+      this.Employee = data;
+    });
   }
 
   removeEmployee(employee, index) {
-    if(window.confirm('Are you sure?')) {
-        this.apiService.deleteEmployee(employee._id).subscribe((data) => {
-          this.Employee.splice(index, 1);
-        }
-      )    
+    if (window.confirm('Are you sure?')) {
+      this.apiService.deleteEmployee(employee._id).subscribe((data) => {
+        this.Employee.splice(index, 1);
+      });
     }
   }
-
 }
