@@ -1,6 +1,6 @@
 export class Employee {
-   name: string;
-   email: string;
-   designation: string;
-   phoneNumber: number;
+  name: string;
+  email: string;
+  designation: string;
+  phoneNumber: number;
 }
